Enabled JsonSerialization object in HTTP methods
Use of JsonSerialization object in Http methods. As System.Text.Json and Newtonsoft.Json usually simply ignore each other CustomAttributes, it result in a standard plain serialization/deserialization with no control over the serialized data. This also means that the behavior of the serializers can change without prior notice. JsonSerialization class simply instruct Restling to use the right serialization library based on the attributes that are present in the classes specified as generics type. As System.Json.Text is provided with dotnet, no compatibility issues will be expected in the near future.